[FLINK-14635][e2e tests] Use non-relocated imports for AWS SDK Kinesis classes in tests
  - Moving pubsub client testutil to main project's to test jar
